+++ src/models/Settings.php
@@ -8,6 +8,7 @@
 namespace craft\commerce\taxjar\models;
 
 use craft\commerce\base\Model;
+use craft\helpers\App;
 
 /**
  * Settings model.
@@ -18,22 +19,71 @@ use craft\commerce\base\Model;
 class Settings extends Model
 {
     /**
-     * @var string
+     * @var ?string
      */
-    public $apiKey;
+    private ?string $_apiKey = null;
 
     /**
-     * @var bool
+     * @var ?string
      */
-    public $useSandbox = false;
+    private ?string $_useSandbox = null;
 
     /**
-     * @inheritdoc
+     * @inerhitdoc
      */
-    public function defineRules(): array
+    public function attributes(): array
     {
-        $rules = parent::defineRules();
+        $names = parent::attributes();
+        $names[] = 'apiKey';
+        $names[] = 'useSandbox';
 
-        return $rules;
+        return $names;
+    }
+
+    public function fields(): array
+    {
+        return [
+            'apiKey' => fn() => $this->getApiKey(false),
+            'useSandbox' => fn() => $this->getUseSandbox(false),
+        ];
+    }
+
+    /**
+     * @param bool $parse
+     * @return string
+     */
+    public function getApiKey(bool $parse = true): string
+    {
+        return ($parse ? App::parseEnv($this->_apiKey) : $this->_apiKey) ?? '';
+    }
+
+    /**
+     * @param string $value
+     * @return void
+     */
+    public function setApiKey(string $value): void
+    {
+        $this->_apiKey = $value;
+    }
+
+    /**
+     * @param bool $parse
+     * @return bool|string
+     */
+    public function getUseSandbox(bool $parse = true): bool|string
+    {
+        if (!$parse) {
+            return $this->_useSandbox ?? false;
+        }
+        return (bool) App::parseBooleanEnv($this->_useSandbox);
+    }
+
+    /**
+     * @param bool|string $value
+     * @return void
+     */
+    public function setUseSandbox(bool|string $value): void
+    {
+        $this->_useSandbox = $value;
     }
 }